add pooling layer tests

+++ Tests/UnitTests/V2LibraryTests/FunctionTests.cpp
@@ -468,11 +468,11 @@ void TestPooling()
             auto ceilOutput = ceilPooling->Output().Shape();
             if (sameOutDim && floorOutput != ceilOutput) {
                 ReportFailure("Ceiling and Flooring expect same output size in the case between [%d * %d * %d] and [%d * %d * %d]", 
-                    floorOutput[0], floorOutput[1], floorOutput[2], ceilOutput[0], ceilOutput[1], ceilOutput[2]);
+                    (int)floorOutput[0], (int)floorOutput[1], (int)floorOutput[2], (int)ceilOutput[0], (int)ceilOutput[1], (int)ceilOutput[2]);
             }
             if (!sameOutDim && floorOutput == ceilOutput) {
                 ReportFailure("Ceiling and Flooring expect different output size in the case between [%d * %d * %d] and [%d * %d * %d]",
-                    floorOutput[0], floorOutput[1], floorOutput[2], ceilOutput[0], ceilOutput[1], ceilOutput[2]);
+                    (int)floorOutput[0], (int)floorOutput[1], (int)floorOutput[2], (int)ceilOutput[0], (int)ceilOutput[1], (int)ceilOutput[2]);
             }
         };
